FilmBase/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
requests
python-dotenv
Werkzeug
pytest
pytest-flask